Update Granular access control

+++ granular_access/lambda_functions/granular_access/granular_access.py
@@ -25,29 +25,58 @@ def describe_user(username, account_id, aws_region):
     res = qs_client.describe_user(
         UserName=username,
         AwsAccountId=account_id,
-        Namespace='gademo'
+        Namespace='default'
     )
     return res
 
-def register_user(username, Email,UserRole, Arn, account_id, aws_region):
+
+def create_namespace(account_id, aws_region, ns):
     qs_client = boto3.client('quicksight', region_name=aws_region)
-    res = qs_client.register_user(
-        IdentityType='IAM',
-        Email=Email,
-        UserRole=UserRole,
-        IamArn=Arn,
-        SessionName=username,
+    res = qs_client.create_namespace(
         AwsAccountId=account_id,
-        Namespace='gademo'
+        Namespace=ns,
+        IdentityStore='QUICKSIGHT'
     )
     return res
+
+
+def register_user(aws_region, Identity, Email, User, AccountId, Arn=None, Session=None, NS='default', Role='READER'):
+    qs_client = boto3.client('quicksight', region_name=aws_region)
+    if Identity == 'QUICKSIGHT':
+        response = qs_client.register_user(
+            IdentityType='QUICKSIGHT',
+            Email=Email,
+            UserRole=Role,
+            AwsAccountId=AccountId,
+            Namespace=NS,
+            UserName=User)
+    elif Identity == 'IAM' and Session is None:
+        response = qs_client.register_user(
+            IdentityType=Identity,
+            Email=Email,
+            UserRole=Role,
+            IamArn=Arn,
+            AwsAccountId=AccountId,
+            Namespace=NS)
+    elif Identity == 'IAM' and Session is not None:
+        response = qs_client.register_user(
+            IdentityType=Identity,
+            Email=Email,
+            UserRole=Role,
+            IamArn=Arn,
+            SessionName=Session,
+            AwsAccountId=AccountId,
+            Namespace=NS)
+
+    return response
+
 
 def delete_user(username, account_id, aws_region):
     qs_client = boto3.client('quicksight', region_name=aws_region)
     res = qs_client.delete_user(
         UserName=username,
         AwsAccountId=account_id,
-        Namespace='gademo'
+        Namespace='default'
     )
     return res
 
@@ -57,7 +86,7 @@ def create_group(userrole, account_id, aws_region):
     res = qs_client.create_group(
         GroupName=userrole,
         AwsAccountId=account_id,
-        Namespace='gademo'
+        Namespace='default'
     )
     return res
 
@@ -68,7 +97,7 @@ def create_group_membership(username, userrole, account_id, aws_region):
         MemberName=username,
         GroupName=userrole,
         AwsAccountId=account_id,
-        Namespace='gademo'
+        Namespace='default'
     )
     return res
 
@@ -79,7 +108,7 @@ def delete_group_membership(username, userrole, account_id, aws_region):
         MemberName=username,
         GroupName=userrole,
         AwsAccountId=account_id,
-        Namespace='gademo'
+        Namespace='default'
     )
     return res
 
@@ -116,7 +145,7 @@ def list_groups(
     return _list(
         func_name="list_groups",
         attr_name="GroupList",
-        Namespace='gademo',
+        Namespace='default',
         account_id=account_id,
         aws_region=aws_region
     )
@@ -134,11 +163,35 @@ def list_dashboards(
     )
 
 
+def list_analyses(
+        account_id,
+        aws_region
+) -> List[Dict[str, Any]]:
+    return _list(
+        func_name="list_analyses",
+        attr_name="AnalysisSummaryList",
+        account_id=account_id,
+        aws_region=aws_region
+    )
+
+
+def list_themes(
+        account_id,
+        aws_region
+) -> List[Dict[str, Any]]:
+    return _list(
+        func_name="list_themes",
+        attr_name="ThemeSummaryList",
+        account_id=account_id,
+        aws_region=aws_region
+    )
+
+
 def list_group_memberships(
         group_name: str,
         account_id: str,
         aws_region: str,
-        namespace: str = "gademo"
+        namespace: str = "default"
 ) -> List[Dict[str, Any]]:
     return _list(
         func_name="list_group_memberships",
@@ -154,7 +207,7 @@ def list_users(account_id, aws_region) -> List[Dict[str, Any]]:
     return _list(
         func_name="list_users",
         attr_name="UserList",
-        Namespace='gademo',
+        Namespace='default',
         account_id=account_id,
         aws_region=aws_region
     )
@@ -164,7 +217,7 @@ def list_user_groups(UserName, account_id, aws_region) -> List[Dict[str, Any]]:
     return _list(
         func_name="list_user_groups",
         attr_name="GroupList",
-        Namespace='gademo',
+        Namespace='default',
         UserName=UserName,
         account_id=account_id,
         aws_region=aws_region
@@ -248,10 +301,12 @@ def lambda_handler(event, context):
             currentmembership[group] = []
 
     users = list_users(account_id, aws_region)
+    # print(users)
     for user in users:
         useralias = user['UserName'].split('/')[-1]
         useralias = useralias.split('@')[0]
         new[useralias] = user['UserName']
+        # print(user['UserName'])
         usergroups = list_user_groups(user['UserName'], account_id, aws_region)
         if len(usergroups) == 0:
             if 'None' in currentmembership:
@@ -273,11 +328,20 @@ def lambda_handler(event, context):
     # build group-members mapping from membership file
     memberships = {}
 
+    # load qs user role information
+    roles = get_ssm_parameters('/qs/config/roles')
+    print(roles)
+
+    # load qs namespace information
+    ns = get_ssm_parameters('/qs/config/ns')
+    print(ns)
+
     # load group mapping information
     with open(path) as csv_file:
         members = csv.reader(csv_file, delimiter=',')
         for member in members:
             alias = member[1]
+            email = member[2]
             # create group
             group = member[0]
             newgroup = 'quicksight-fed-' + group.lower()
@@ -297,7 +361,30 @@ def lambda_handler(event, context):
 
             # handle every user
             if alias:
-                # add user into the group
+                # register user is user is new
+                if alias not in users:
+                    try:
+                        response = register_user(aws_region, 'IAM', email, alias, account_id,
+                                                 Arn='arn:aws:iam::' + account_id + ':role/quicksight-fed-us-users',
+                                                 Session=email, NS=ns[group], Role=roles[group])
+                        qs_usr_name = 'quicksight-fed-us-users/' + email
+                        print(qs_usr_name + " is registered")
+                    except Exception as e:
+                        print(e)
+                        pass
+
+                    try:
+                        response = create_group_membership(qs_usr_name, newgroup, account_id, aws_region)
+                        print(qs_usr_name + " is added into " + newgroup)
+
+                    except Exception as e:
+                        if str(e).find('does not exist'):
+                            # print(e)
+                            pass
+                        else:
+                            raise e
+
+                # add user into the group if user exists
                 if alias in users:
                     qs_usr_name = users[alias]
                     email = users[alias].split('/')[-1]
@@ -315,15 +402,16 @@ def lambda_handler(event, context):
                                 pass
                             else:
                                 raise e
-                        if newgroup in ['quicksight-fed-bi-admin']:
-                            print("add new admin")
+                        # if newgroup in ['quicksight-fed-bi-admin']:
+                        if group in roles:
+                            print("update role")
                             try:
                                 qs_client.update_user(
                                     UserName=qs_usr_name,
                                     AwsAccountId=account_id,
-                                    Namespace='gademo',
+                                    Namespace=ns[group],
                                     Email=email,
-                                    Role='ADMIN'
+                                    Role=roles[group]
                                 )
                             except Exception as e:
                                 if str(e).find('does not exist'):
@@ -331,7 +419,7 @@ def lambda_handler(event, context):
                                     pass
                                 else:
                                     raise e
-                        elif newgroup in ['quicksight-fed-bi-developer',
+                        """elif newgroup in ['quicksight-fed-bi-developer',
                                           'quicksight-fed-power-reader']:
                             print("add new dev")
                             if qs_usr_name not in memberships['quicksight-fed-bi-admin']:
@@ -339,7 +427,7 @@ def lambda_handler(event, context):
                                     qs_client.update_user(
                                         UserName=qs_usr_name,
                                         AwsAccountId=account_id,
-                                        Namespace='gademo',
+                                        Namespace='default',
                                         Email=email,
                                         Role='AUTHOR'
                                     )
@@ -348,13 +436,13 @@ def lambda_handler(event, context):
                                         # print(e)
                                         pass
                                     else:
-                                        raise e
+                                        raise e"""
 
                 # write current group-users mapping into a dict "membership"
                 if newgroup in memberships:
-                    if 'qs_usr_name' in locals(): # User in membership.csv already registered as a QS user
+                    if 'qs_usr_name' in locals():  # User in membership.csv already registered as a QS user
                         memberships[newgroup].append(qs_usr_name)
-                    else: # User in membership.csv do not register as a QS user yet
+                    else:  # User in membership.csv do not register as a QS user yet
                         memberships[newgroup].append(alias)
 
                 else:
@@ -374,7 +462,7 @@ def lambda_handler(event, context):
         # print(groups)
         if len(groups) == 0:
             # currentmembership['None'] = user['UserName']
-            #if user['UserName'].split("/", 1)[0] != 'i-app-us-east-1-quicksight-admin':
+            if user['UserName'].split("/", 1)[0] != 'Administrator':
                 delete_user(user['UserName'], account_id, aws_region)
         else:
             for group in groups:
@@ -387,15 +475,18 @@ def lambda_handler(event, context):
 
     # update access permissions
     dashboards = list_dashboards(account_id, lambda_aws_region)
+    analyses = list_analyses(account_id, lambda_aws_region)
     datasets = list_datasets(account_id, lambda_aws_region)
     datasources = list_data_sources(account_id, lambda_aws_region)
+    themes = list_themes(account_id, lambda_aws_region)
     permissions = get_ssm_parameters('/qs/config/access')
     print(permissions)
     permissions = permissions['Permissions']
     reportlist = []
     for permission in permissions:
-        arn = "arn:aws:quicksight:" + aws_region + ":" + account_id + ":group/gademo/quicksight-fed-" + permission[
-            'Group_Name'].lower()
+        arn = 'arn:aws:quicksight:' + aws_region + ':' + account_id + ":group/" + ns[
+            permission['Group_Name']] + "/quicksight-fed-" + permission[
+                  'Group_Name'].lower()
         reportnamels = permission['Reports']
         print(reportnamels)
         if len(reportnamels) > 0:
@@ -451,7 +542,10 @@ def lambda_handler(event, context):
                             )
 
                         except Exception as e:
-                            print(e)
+                            if str(e).find('FILE'):
+                                pass
+                            else:
+                                print(e)
 
                 for datasource in datasources:
                     datasourceid = datasource['DataSourceId']
@@ -473,7 +567,64 @@ def lambda_handler(event, context):
                         )
 
                     except Exception as e:
-                        print(e)
+                        if str(e).find('FILE'):
+                            pass
+                        else:
+                            print(e)
+
+                for analysis in analyses:
+                    if analysis['Status'] != 'DELETED':
+                        analysisid = analysis['AnalysisId']
+                        try:
+                            response = qs_local_client.update_analysis_permissions(
+                                AwsAccountId=account_id,
+                                AnalysisId=analysisid,
+                                GrantPermissions=[
+                                    {
+                                        'Principal': arn,
+                                        'Actions': ['quicksight:RestoreAnalysis',
+                                                    'quicksight:UpdateAnalysisPermissions',
+                                                    'quicksight:DeleteAnalysis',
+                                                    'quicksight:QueryAnalysis',
+                                                    'quicksight:DescribeAnalysisPermissions',
+                                                    'quicksight:DescribeAnalysis',
+                                                    'quicksight:UpdateAnalysis']
+                                    },
+                                ]
+                            )
+
+                        except Exception as e:
+                            print(e)
+
+                for theme in themes:
+                    if theme['ThemeId'] not in ['SEASIDE', 'CLASSIC', 'MIDNIGHT']:
+                        themeid = theme['ThemeId']
+                        try:
+                            response = qs_local_client.update_theme_permissions(
+                                AwsAccountId=account_id,
+                                ThemeId=themeid,
+                                GrantPermissions=[
+                                    {
+                                        'Principal': arn,
+                                        'Actions': ["quicksight:DescribeTheme",
+                                                    "quicksight:DescribeThemeAlias",
+                                                    "quicksight:ListThemeAliases",
+                                                    "quicksight:ListThemeVersions",
+                                                    "quicksight:DeleteTheme",
+                                                    "quicksight:UpdateTheme",
+                                                    "quicksight:CreateThemeAlias",
+                                                    "quicksight:DeleteThemeAlias",
+                                                    "quicksight:UpdateThemeAlias",
+                                                    "quicksight:UpdateThemePermissions",
+                                                    "quicksight:DescribeThemePermissions"
+                                                    ]
+                                    },
+                                ]
+                            )
+
+                        except Exception as e:
+                            print(e)
+
             elif reportnamels[0] == 'read-all':
                 for dashboard in dashboards:
                     dashboardid = dashboard['DashboardId']
@@ -520,7 +671,10 @@ def lambda_handler(event, context):
                             )
 
                         except Exception as e:
-                            print(e)
+                            if str(e).find('FILE'):
+                                pass
+                            else:
+                                print(e)
 
                 for datasource in datasources:
                     datasourceid = datasource['DataSourceId']
@@ -539,7 +693,10 @@ def lambda_handler(event, context):
                         )
 
                     except Exception as e:
-                        print(e)
+                        if str(e).find('FILE'):
+                            pass
+                        else:
+                            print(e)
             else:
                 for reportname in reportnamels:
                     ids = get_dashboard_ids(reportname, account_id, lambda_aws_region)
@@ -569,10 +726,10 @@ def lambda_handler(event, context):
         for dashboard in dashboards:
             dashboardid = dashboard['DashboardId']
             if dashboardid not in reportlist:
-                print(dashboardid)
+                # print(dashboardid)
                 if group not in ['quicksight-fed-bi-developer', 'quicksight-fed-bi-admin',
-                                'quicksight-fed-power-reader']:
-                    print("revoke " + group)
+                                 'quicksight-fed-power-reader']:
+                    print("revoke " + group + "dashboard id: " + dashboardid + "(name: " + dashboard['Name'] + ")")
                     try:
                         response = qs_local_client.update_dashboard_permissions(
                             AwsAccountId=account_id,
@@ -593,7 +750,7 @@ def lambda_handler(event, context):
                             raise e
 
     # get current quicksight groups
-    groups = list_groups(account_id, aws_region)
+    """groups = list_groups(account_id, aws_region)
     new = []
     for group in groups:
         new.append(group['GroupName'])
@@ -610,6 +767,7 @@ def lambda_handler(event, context):
     users = list_users(account_id, aws_region)
     for user in users:
         new.append(user['UserName'])
+        print(user['UserName'])
         usergroups = list_user_groups(user['UserName'], account_id, aws_region)
         if len(usergroups) == 0:
             if 'None' in currentmembership:
@@ -626,4 +784,4 @@ def lambda_handler(event, context):
                     currentmembership[group['GroupName']] = []
                     currentmembership[group['GroupName']].append(user['UserName'])
     print("here it is current users mapping:")
-    print(currentmembership)
+    print(currentmembership)"""
